requirements.txt
# pip-requirements
mcp[cli]>=1.6.0
flask>=2.2.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
httpx>=0.28.1
psutil>=5.9.0
aiohttp>=3.8.4
openai>=0.28.0
